Make the remaining defines.h constants constexpr
Every value macro left in defines.h is now an inline constexpr with a
Google-style name, and every use is renamed. The preprocessor feature
flags that are tested with #if (RELEASE_VERSION, TIMING_FIX, the
commented DEV_VERSION and DONGLE) stay as they are.

Types follow the use sites rather than the old literal suffixes:

- The command-line keywords kParmPlaytest and kParmInstall are uint32_t
  hashes. Obfuscate() now returns uint32_t too, and the CheatCodes,
  PlayCodes and EditorCodes tables in const.h drop their long casts, so
  the switch in Parse_Command_Line compares like with like on every
  platform instead of relying on long being 64 bits.
- kStructFlag*, kUnitFlag* and kInfantryFlagDog are uint64_t: StructType
  has 90 entries and the flagged ones reach bit 34, which the old
  `1L << n` only survived on LP64.
- kHouseFlag* are int (20 houses), matching ThemeControl::Owner and
  Get_Ownable(). kTheaterFlag* and kZoneFlag* are int masks.
- kTpfButton, kTpfEButton and kTpfText are TextPrintType; kThreatGround
  is ThreatType; kSpcChrono2 is SpecialWeaponType; kTargetNone is a
  value-initialised TARGET as in Tiberian Dawn.
- kRefreshEol and kRefreshSidebar are int16_t, the element type of the
  occupy and overlap lists they terminate.
- Buffer sizes, disk-space thresholds, region geometry, the mantissa and
  exponent widths of TARGET and the kMax* limits are int. The file names
  are constexpr char arrays.

Dropped as dead: NETWORK, FOREIGN_VERSION and FOREIGN_VERSION_NUMBER
(never tested), the VESSELF_ and AIRCRAFTF_ families, the UNITF_ entries
other than HARVESTER and MCV, and the STRUCTF_ entries nothing reads.
The remaining AIRCRAFTF_ mention sits in an #ifdef NEVER block in
house.cc alongside Tiberian Dawn names that never existed here.

DIR_SW_X1 and DIR_SW_X2 were macros shadowing enumerators that existed
only in the #ifdef NEVER copy of DirType in defines.h. They are now
enumerators of the live DirType in face.h.

kMapRegionHeight rounds with kRegionHeight; the old macro used
REGION_WIDTH in that term. Both are 4, so the value is unchanged.

+++ src/ra/const.h
@@ -1,6 +1,7 @@
 #ifndef CNC_RED_ALERT_RA_CONST_H_
 #define CNC_RED_ALERT_RA_CONST_H_
 
+#include <cstdint>
 #include <string>
 
 #include "ra/config.h"
@@ -17,31 +18,31 @@ extern GroundType Ground[LAND_COUNT];
 **	These are the access passwords used to activate cheat mode, editor mode,
 **	and special game options.
 */
-constexpr long PlayCodes[] = {static_cast<long>(0xE0792D6D),  // Dwight Okahara
-                              static_cast<long>(0x90046ECF),  // Paul S. Mudra
-                              static_cast<long>(0xC3EE9A26),  // Frank Klepaki
-                              static_cast<long>(0xED382178),  // Ed Del Castillo
-                              0L};
+constexpr uint32_t PlayCodes[] = {0xE0792D6D,  // Dwight Okahara
+                                  0x90046ECF,  // Paul S. Mudra
+                                  0xC3EE9A26,  // Frank Klepaki
+                                  0xED382178,  // Ed Del Castillo
+                                  0};
 
-constexpr long CheatCodes[] = {static_cast<long>(0xA0E2AB53),  // Joseph Hewitt
-                               static_cast<long>(0x00532693),  // Mike Lightner
-                               static_cast<long>(0x7DDFF824),  // Joe Bostic
-                               static_cast<long>(0x2CB5CF01),  // Phil Gorrow
-                               static_cast<long>(0xB5B63531),  // Bill Randolph
-                               static_cast<long>(0xDFABC23A),  // Adam Isgreen
-                               static_cast<long>(0x52B19A22),  // Erik Yeo
-                               static_cast<long>(0xBE79088C),  // David Dettmer
-                               static_cast<long>(0xB216AE7E),  // Barry Green
-                               static_cast<long>(0x0E07B213),  // Steve Tall
-                               0L};
+constexpr uint32_t CheatCodes[] = {0xA0E2AB53,  // Joseph Hewitt
+                                   0x00532693,  // Mike Lightner
+                                   0x7DDFF824,  // Joe Bostic
+                                   0x2CB5CF01,  // Phil Gorrow
+                                   0xB5B63531,  // Bill Randolph
+                                   0xDFABC23A,  // Adam Isgreen
+                                   0x52B19A22,  // Erik Yeo
+                                   0xBE79088C,  // David Dettmer
+                                   0xB216AE7E,  // Barry Green
+                                   0x0E07B213,  // Steve Tall
+                                   0};
 
-constexpr long EditorCodes[] = {static_cast<long>(0xA2C09326),  // Erik Yeo
-                                static_cast<long>(0x1F944BB3),  // Mike Lightner
-                                static_cast<long>(0xDE07154D),  // Adam Isgreen
-                                static_cast<long>(0x0E07B213),  // Steve Tall
-                                static_cast<long>(0x16B170B1),  // Joe Bostic
-                                // static_cast<long>(0x7DDFF824),  // PASSWORD
-                                0L};
+constexpr uint32_t EditorCodes[] = {0xA2C09326,  // Erik Yeo
+                                    0x1F944BB3,  // Mike Lightner
+                                    0xDE07154D,  // Adam Isgreen
+                                    0x0E07B213,  // Steve Tall
+                                    0x16B170B1,  // Joe Bostic
+                                    // 0x7DDFF824,  // PASSWORD
+                                    0};
 
 /***********************************************************************************************
 **	Unit order names. These names correspond to the player selectable orders
