+++ src/ra/init.h
@@ -1,8 +1,10 @@
 #ifndef CNC_RED_ALERT_RA_INIT_H_
 #define CNC_RED_ALERT_RA_INIT_H_
 
+#include <cstdint>
+
 void Load_Title_Page(bool visible = false);
-long Obfuscate(const char* string);
+uint32_t Obfuscate(const char* string);
 void Anim_Init();
 bool Init_Game(int argc, char* argv[]);
 bool Select_Game(bool fade = false);
